+++ src/Post/constants/index.ts
@@ -1,0 +1,8 @@
+import { Option } from '../types';
+
+export const LIMIT_PER_PAGE = 10;
+
+export const SELECT_OPTIONS: Option[] = [
+  { value: 'title', name: 'By title' },
+  { value: 'body', name: 'By description' },
+];
