--- src/modules/Post/pages/PostIdPage.tsx
@@ -1,38 +1,0 @@
-import React, { useEffect, useState } from 'react';
-import { useParams } from 'react-router-dom';
-import Loader from '../../../components/UI/Loader/Loader';
-import { useFetching } from '../../../hooks/useFetching';
-import { Post } from '../types';
-import { getPostById } from '../services/PostService';
-
-const PostIdPage: React.FC = () => {
-  const params = useParams();
-  const [post, setPost] = useState<Post | null>(null);
-  const [fetchPostById, isLoading] = useFetching(async (id: string) => {
-    const response = await getPostById(id);
-    setPost(response.data);
-  });
-
-  useEffect(() => {
-    fetchPostById(params.id);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [params.id]);
-
-  return (
-    <div>
-      <h1>Page post ID = {params.id}</h1>
-
-      {isLoading ? (
-        <Loader />
-      ) : (
-        post && (
-          <div>
-            {post.id}. {post.title}
-          </div>
-        )
-      )}
-    </div>
-  );
-};
-
-export default PostIdPage;
