--- src/pages/About.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const About: React.FC = () => {
-  return <h1>About my react app</h1>;
-};
-
-export default About;
